unit test fix

+++ src/conf.hpp
@@ -25,7 +25,7 @@ class Conf {
   uint16_t working_threads_num_ = 5;
 
   // cache
-  int cache_size_bytes_ = 1 << 25;
+  size_t cache_size_bytes_ = 1 << 25;
   int cache_default_max_age_ = 300;
   int cache_clean_interval_ = 10;
 
@@ -47,7 +47,7 @@ class Conf {
   void set_lua_config_path(std::string path) { lua_config_path_ = path; }
   sol::state& get_lua() { return lua_; }
 
-  int get_cache_size_bytes() { return cache_size_bytes_; }
+  size_t get_cache_size_bytes() { return cache_size_bytes_; }
   uint16_t get_port() { return port_; }
   uint16_t get_working_threads_num() { return working_threads_num_; }
   OriginType get_origin_type() { return origin_type_; }
